feat: created state types
