+++ src/validate.ts
@@ -1,4 +1,4 @@
-import { State } from "./index.js";
+import { GenericInitialState, State } from "./state.js";
 
 export class InvalidStatesError extends Error {}
 export class InvalidInitialStateError extends Error {}
@@ -13,7 +13,7 @@ export function validateStates<TContext, TEvent>(
 			"States must be an array of at least 2 elements",
 		);
 
-	const initial = states.filter((s) => s.initial);
+	const initial = states.filter((s) => (s as GenericInitialState).initial);
 	if (initial.length !== 1) {
 		throw new InvalidInitialStateError(
 			"There must be one and only initial state",
@@ -29,7 +29,10 @@ export function validateStates<TContext, TEvent>(
 		throw new InvalidStateIdError("Ids must be unique");
 	}
 
-	if (states.some((s) => s.autoTransition && s.transitionGuard)) {
+	if (
+		// biome-ignore lint/suspicious/noExplicitAny: <during validation we can have unexpected types, that do not match existing>
+		states.some((s: any) => s.autoTransition && s.transitionGuard)
+	) {
 		throw new InvalidTransitionCondition(
 			"State with autoTransition cannot have property transitionGuard",
 		);
@@ -37,7 +40,8 @@ export function validateStates<TContext, TEvent>(
 
 	if (
 		states.some(
-			(s) =>
+			// biome-ignore lint/suspicious/noExplicitAny: <during validation we can have unexpected types, that do not match existing>
+			(s: any) =>
 				(!s.final && !s.transitionTo) ||
 				(s.final && (s.autoTransition || s.transitionTo || s.transitionGuard)),
 		)
@@ -46,8 +50,8 @@ export function validateStates<TContext, TEvent>(
 			"State must have autoTransition or transitionTo if not final",
 		);
 	}
-
-	if (states.some((s) => !s.final && s.onFinal)) {
+	// biome-ignore lint/suspicious/noExplicitAny: <during validation we can have unexpected types, that do not match existing>
+	if (states.some((s: any) => !s.final && s.onFinal)) {
 		throw new InvalidTransitionCondition(
 			"State that are not final cannot have property onFinal",
 		);
